requirements.txt
clickhouse-driver[lz4]
apache-airflow==2.1.2